split chart block tests into department classes
